Aligning with Charts 4 package sturucture
