+++ addon/src/main/java/com/vaadin/addon/charts/Chart.java
@@ -16,31 +16,6 @@
  */
 package com.vaadin.addon.charts;
 import com.vaadin.addon.charts.events.ConfigurationChangeListener;
-import com.vaadin.addon.charts.events.domevents.ChartAddSeriesEvent;
-import com.vaadin.addon.charts.events.domevents.ChartAfterPrintEvent;
-import com.vaadin.addon.charts.events.domevents.ChartBeforePrintEvent;
-import com.vaadin.addon.charts.events.domevents.ChartClickEvent;
-import com.vaadin.addon.charts.events.domevents.ChartDrillupAllEvent;
-import com.vaadin.addon.charts.events.domevents.ChartDrillupEvent;
-import com.vaadin.addon.charts.events.domevents.ChartLoadEvent;
-import com.vaadin.addon.charts.events.domevents.ChartRedrawEvent;
-import com.vaadin.addon.charts.events.domevents.ChartSelectionEvent;
-import com.vaadin.addon.charts.events.domevents.SeriesCheckboxClickEvent;
-import com.vaadin.addon.charts.events.domevents.DrilldownEvent;
-import com.vaadin.addon.charts.events.domevents.SeriesLegendItemClickEvent;
-import com.vaadin.addon.charts.events.domevents.PointClickEvent;
-import com.vaadin.addon.charts.events.domevents.PointMouseOutEvent;
-import com.vaadin.addon.charts.events.domevents.PointMouseOverEvent;
-import com.vaadin.addon.charts.events.domevents.PointRemoveEvent;
-import com.vaadin.addon.charts.events.domevents.PointSelectEvent;
-import com.vaadin.addon.charts.events.domevents.PointUnselectEvent;
-import com.vaadin.addon.charts.events.domevents.PointUpdateEvent;
-import com.vaadin.addon.charts.events.domevents.SeriesAfterAnimateEvent;
-import com.vaadin.addon.charts.events.domevents.SeriesClickEvent;
-import com.vaadin.addon.charts.events.domevents.SeriesHideEvent;
-import com.vaadin.addon.charts.events.domevents.SeriesMouseOutEvent;
-import com.vaadin.addon.charts.events.domevents.SeriesMouseOverEvent;
-import com.vaadin.addon.charts.events.domevents.SeriesShowEvent;
 import com.vaadin.addon.charts.model.ChartType;
 import com.vaadin.addon.charts.model.Configuration;
 import com.vaadin.addon.charts.model.DataSeries;
