+++ addon/src/main/java/com/vaadin/addon/charts/ChartRedrawEvent.java
@@ -1,4 +1,4 @@
-package com.vaadin.addon.charts.events.domevents;
+package com.vaadin.addon.charts;
 
 import com.vaadin.addon.charts.Chart;
 import com.vaadin.ui.event.ComponentEvent;
