+++ addon/src/main/java/com/vaadin/addon/charts/PointClickEvent.java
@@ -1,7 +1,5 @@
-package com.vaadin.addon.charts.events.domevents;
+package com.vaadin.addon.charts;
 
-import com.vaadin.addon.charts.Chart;
-import com.vaadin.addon.charts.events.domevents.details.MouseEventDetails;
 import com.vaadin.ui.event.ComponentEvent;
 import com.vaadin.ui.event.DomEvent;
 import com.vaadin.ui.event.EventData;
