+++ addon/src/main/java/com/vaadin/addon/charts/SeriesCheckboxClickEvent.java
@@ -1,4 +1,4 @@
-package com.vaadin.addon.charts.events.domevents;
+package com.vaadin.addon.charts;
 
 /*
  * #%L
@@ -17,7 +17,6 @@ package com.vaadin.addon.charts.events.domevents;
  * #L%
  */
 
-import com.vaadin.addon.charts.Chart;
 import com.vaadin.ui.event.ComponentEvent;
 import com.vaadin.ui.event.DomEvent;
 import com.vaadin.ui.event.EventData;
